Upgrade to accepting transaction factory
The transaction factory is moved to the visitor namespace and now
returns an accepting transaction factory.
