+++ src/Visitor/TransactionFactory.php
@@ -1,0 +1,26 @@
+<?php
+/**
+ * Copyright MediaCT. All rights reserved.
+ * https://www.mediact.nl
+ */
+
+namespace Johmanx10\Transaction\Visitor;
+
+use Johmanx10\Transaction\OperationInterface;
+use Johmanx10\Transaction\Transaction;
+
+class TransactionFactory implements TransactionFactoryInterface
+{
+    /**
+     * Create a transaction for the given operations.
+     *
+     * @param OperationInterface ...$operations
+     *
+     * @return AcceptingTransactionInterface
+     */
+    public function createTransaction(
+        OperationInterface ...$operations
+    ): AcceptingTransactionInterface {
+        return new Transaction(...$operations);
+    }
+}
