+++ src/Visitor/TransactionFactoryInterface.php
@@ -4,20 +4,20 @@
  * https://www.mediact.nl
  */
 
-namespace Johmanx10\Transaction;
+namespace Johmanx10\Transaction\Visitor;
 
-class TransactionFactory implements TransactionFactoryInterface
+use Johmanx10\Transaction\OperationInterface;
+
+interface TransactionFactoryInterface
 {
     /**
      * Create a transaction for the given operations.
      *
      * @param OperationInterface ...$operations
      *
-     * @return TransactionInterface
+     * @return AcceptingTransactionInterface
      */
     public function createTransaction(
         OperationInterface ...$operations
-    ): TransactionInterface {
-        return new Transaction(...$operations);
-    }
+    ): AcceptingTransactionInterface;
 }
